+++ app/helpers/application_helper.rb
@@ -123,9 +123,8 @@ module ApplicationHelper
     closed = issue.closed? ? content_tag(:span, l(:label_closed_issues), :class => "hidden-for-sighted") : ""
     s = ActiveSupport::SafeBuffer.new
     s << "#{issue.project} - " if options[:project]
-    s << link_to("#{closed}#{h(options[:before_text].to_s)}#{h(issue.tracker)} ##{issue.id}".html_safe,
-                issue,
-                :class => issue.css_classes,
+    s << link_to("#{closed}#{h(options[:before_text].to_s)}#{(issue.kind.nil?) ? '' : h(issue.kind.name)} ##{issue.id}".html_safe,
+                work_package_path(issue),
                 :title => h(title))
     s << ": #{subject}" if subject
     s
@@ -714,7 +713,7 @@ module ApplicationHelper
           case prefix
           when nil
             if issue = Issue.visible.find_by_id(oid, :include => :status)
-              link = link_to("##{oid}", {:only_path => only_path, :controller => '/issues', :action => 'show', :id => oid},
+              link = link_to("##{oid}", work_package_path(:id => oid, :only_path => only_path),
                                         :class => issue.css_classes,
                                         :title => "#{truncate(issue.subject, :length => 100)} (#{issue.status.name})")
             end
@@ -1144,18 +1143,18 @@ module ApplicationHelper
     end
 
     link_to(h(text),
-            project_planning_element_path(planning_element.project, planning_element),
+            work_package_path(planning_element),
             :title => planning_element.subject)
   end
 
   def planning_element_quick_info(planning_element)
     start_date_change = ""
-    end_date_change = ""
+    due_date_change = ""
 
     journals = planning_element.journals.find(:all, :conditions => ["created_at >= ?", Date.today.to_time - 7.day], :order => "created_at desc")
 
     journals.each do |journal|
-      break if !start_date_change.empty? and !end_date_change.empty?
+      break if !start_date_change.empty? and !due_date_change.empty?
 
       if start_date_change.empty?
         unless journal.changes["start_date"].nil?
@@ -1165,21 +1164,21 @@ module ApplicationHelper
         end
       end
 
-      if end_date_change.empty?
-        unless journal.changes["end_date"].nil?
-          unless journal.changes["end_date"].first.nil?
-            end_date_change = " (<del>#{journal.changes["end_date"].first}</del>)"
+      if due_date_change.empty?
+        unless journal.changes["due_date"].nil?
+          unless journal.changes["due_date"].first.nil?
+            due_date_change = " (<del>#{journal.changes["due_date"].first}</del>)"
           end
         end
       end
     end
 
     link = link_to(h("*#{planning_element.id} #{planning_element.planning_element_status.nil? ? "" : planning_element.planning_element_status.name + ":"} #{planning_element.subject} "),
-                   project_planning_element_path(planning_element.project, planning_element),
+                   work_package_path(planning_element),
                    :title => h("#{truncate(planning_element.subject, :length => 100)} #{planning_element.planning_element_status.nil? ? "" :
                                "(" + planning_element.planning_element_status.name + ")"}"))
-    link += "#{planning_element.start_date.nil? ? "[?]" : planning_element.start_date.to_s}#{start_date_change} – #{planning_element.end_date.nil? ? "[?]" :
-      planning_element.end_date.to_s}#{end_date_change}"
+    link += "#{planning_element.start_date.nil? ? "[?]" : planning_element.start_date.to_s}#{start_date_change} – #{planning_element.due_date.nil? ? "[?]" :
+      planning_element.due_date.to_s}#{due_date_change}"
     link += "#{planning_element.responsible.nil? ? "" : h(" (#{planning_element.responsible.to_s})")}"
 
     return link