+++ spec/models/permitted_params_spec.rb
@@ -15,6 +15,31 @@ describe PermittedParams do
   let(:user) { FactoryGirl.build(:user) }
   let(:admin) { FactoryGirl.build(:admin) }
 
+  describe :permit do
+    it "adds an attribute to be permitted later" do
+      # just taking project_type here as an example, could be anything
+
+      # taking the originally whitelisted params to be restored later
+      original_whitelisted = PermittedParams.instance_variable_get(:@whitelisted_params)
+
+
+      params = ActionController::Parameters.new(:project_type => { "blubs1" => "blubs" } )
+
+      PermittedParams.new(params, user).project_type.should == {}
+
+      PermittedParams.permit(:project_type, :blubs1)
+
+      PermittedParams.new(params, user).project_type.should == { "blubs1" => "blubs" }
+
+
+      PermittedParams.instance_variable_set(:@whitelisted_params, original_whitelisted)
+    end
+
+    it "raises an argument error if key does not exist" do
+      expect{ PermittedParams.permit(:bogus_key) }.to raise_error ArgumentError
+    end
+  end
+
   describe :project_type do
     it "should return name" do
       params = ActionController::Parameters.new(:project_type => { "name" => "blubs" } )
@@ -166,8 +191,8 @@ describe PermittedParams do
       PermittedParams.new(params, user).planning_element.should == hash
     end
 
-    it "should permit end_date" do
-      hash = { "end_date" => "2012-12-12" }
+    it "should permit due_date" do
+      hash = { "due_date" => "2012-12-12" }
 
       params = ActionController::Parameters.new(:planning_element => hash)
 
@@ -175,7 +200,7 @@ describe PermittedParams do
     end
 
     it "should permit scenarios" do
-      hash = { "scenarios" => {'id' => "1", 'start_date' => '2012-01-01', 'end_date' => '2012-01-03' } }
+      hash = { "scenarios" => {'id' => "1", 'start_date' => '2012-01-01', 'due_date' => '2012-01-03' } }
 
       params = ActionController::Parameters.new(:planning_element => hash)
 
@@ -222,12 +247,167 @@ describe PermittedParams do
       PermittedParams.new(params, user).planning_element.should == hash
     end
 
+
     it "should permit responsible_id" do
       hash = { "responsible_id" => "1" }
 
       params = ActionController::Parameters.new(:planning_element => hash)
 
       PermittedParams.new(params, user).planning_element.should == hash
+    end
+  end
+
+  describe :new_work_package do
+    it "should permit subject" do
+      hash = { "subject" => "blubs" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit description" do
+      hash = { "description" => "blubs" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit start_date" do
+      hash = { "start_date" => "2013-07-08" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit due_date" do
+      hash = { "due_date" => "2013-07-08" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit assigned_to_id" do
+      hash = { "assigned_to_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit responsible_id" do
+      hash = { "responsible_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit tracker_id" do
+      hash = { "tracker_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit planning_element_type_id" do
+      hash = { "planning_element_type_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit prioritiy_id" do
+      hash = { "priority_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit parent_issue_id" do
+      hash = { "parent_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit parent_issue_id" do
+      hash = { "parent_issue_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit fixed_version_id" do
+      hash = { "fixed_version_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit estimated_hours" do
+      hash = { "estimated_hours" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit done_ratio" do
+      hash = { "done_ratio" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit status_id" do
+      hash = { "status_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit category_id" do
+      hash = { "category_id" => "1" }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package.should == hash
+    end
+
+    it "should permit watcher_user_ids when the user is allowed to add watchers" do
+      project = double('project')
+
+      user.stub!(:allowed_to?).with(:add_work_package_watchers, project).and_return(true)
+
+      hash = { "watcher_user_ids" => ["1", "2"] }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package(:project => project).should == hash
+    end
+
+    it "should not return watcher_user_ids when the user is not allowed to add watchers" do
+      project = double('project')
+
+      user.stub!(:allowed_to?).with(:add_work_package_watchers, project).and_return(false)
+
+      hash = { "watcher_user_ids" => ["1", "2"] }
+
+      params = ActionController::Parameters.new(:work_package => hash)
+
+      PermittedParams.new(params, user).new_work_package(:project => project).should == {}
     end
   end
 
@@ -256,7 +436,7 @@ describe PermittedParams do
         hash = { field => 'test' }
         params = ActionController::Parameters.new(:user => hash)
 
-        PermittedParams.new(params, admin).user_update_as_admin.should == 
+        PermittedParams.new(params, admin).user_update_as_admin.should ==
           { field => 'test' }
       end
     end