+++ spec/support/flash.rb
@@ -1,0 +1,16 @@
+#-- encoding: UTF-8
+#-- copyright
+# OpenProject is a project management system.
+#
+# Copyright (C) 2012-2013 the OpenProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# See doc/COPYRIGHT.rdoc for more details.
+#++
+
+def disable_flash_sweep
+  controller.instance_eval { flash.stub!(:sweep) }
+end
+